Started writing an actual problem series
Need to do some additional work here. I think that I also probably
want to have a way to define an actual series object, and may want
to consider having a well-known function defined in a "series" module
so that we can get the series that are in there for display. I'll
also need a way to get the generated data for another challenge in the
series for what I'm hoping to do here with a chain of back-to-back
challenges that are all using the same input, but doing different
things with it.