More login updates
I still have work to do here, but it turns out that the version of python
on the netbook is too old to have the threading http server. :/ So I need
to put things in place here and then try this on my desktop where I have
all of the required newer stuff.